feat: more progress on processing

+++ src/custom/OutputPuzzlerPro.java
@@ -4,10 +4,10 @@ import java.util.*;
 
 public class OutputPuzzlerPro {
     public char[][] filledBoardSolution;
-    public int executionTime;
+    public long executionTime;
     public int casesExplored;
 
-    public OutputPuzzlerPro(char[][] filledBoardSolution, int executionTime, int casesExplored) {
+    public OutputPuzzlerPro(char[][] filledBoardSolution, long executionTime, int casesExplored) {
         this.filledBoardSolution = filledBoardSolution;
         this.executionTime = executionTime;
         this.casesExplored = casesExplored;
